[api] Split metal rough material into base class for all PBR materials
