+++ src/3d/CMakeLists.txt
@@ -92,6 +92,7 @@ set(QGIS_3D_SRCS
   materials/qgsmetalroughmaterial3dhandler.cpp
   materials/qgsmetalroughtexturedmaterial3dhandler.cpp
   materials/qgsnullmaterial3dhandler.cpp
+  materials/qgspbrmaterial.cpp
   materials/qgsphongmaterial.cpp
   materials/qgsphongmaterial3dhandler.cpp
   materials/qgsphongtexturedmaterial.cpp
@@ -241,6 +242,7 @@ set(QGIS_3D_HDRS
   materials/qgsmetalroughmaterial3dhandler.h
   materials/qgsmetalroughtexturedmaterial3dhandler.h
   materials/qgsnullmaterial3dhandler.h
+  materials/qgspbrmaterial.h
   materials/qgsphongmaterial.h
   materials/qgsphongmaterial3dhandler.h
   materials/qgsphongtexturedmaterial.h
