+++ src/3d/shaders/pbrDataDefined.vert
